requirements.txt
alembic==1.15.2
boto3==1.37.34
fastapi==0.115.12
firebase-admin==6.8.0
httpx==0.28.1
psycopg[binary]==3.2.9
PyJWT==2.10.1
pydantic-settings==2.9.1
python-multipart==0.0.20
SQLAlchemy==2.0.40
uvicorn[standard]==0.34.2
pytest==8.3.5
